Stop using ESM in extensions
