+++ src/effect/clip_shadow_effect.ts
@@ -3,19 +3,27 @@ import * as GObject                from '@gi/GObject'
 import { SnippetHook, GLSLEffect } from '@gi/Shell'
 
 // local modules
-import { loadShader }              from '../utils/io'
+import { loadShader }              from '@me/utils/io'
+
+// types
+import { Me }                      from '@global'
 
 // ------------------------------------------------------------------- [imports]
 
 const { declarations, code } = loadShader (
-    import.meta.url,
-    './shader/clip_shadow.frag'
+    `${Me.path}/effect/shader/clip_shadow.frag`
 )
 
-class ClipShadowEffect extends GLSLEffect {
-    vfunc_build_pipeline (): void {
-        this.add_glsl_snippet (SnippetHook.FRAGMENT, declarations, code, false)
+export const ClipShadowEffect = GObject.registerClass (
+    {},
+    class extends GLSLEffect {
+        vfunc_build_pipeline (): void {
+            this.add_glsl_snippet (
+                SnippetHook.FRAGMENT,
+                declarations,
+                code,
+                false
+            )
+        }
     }
-}
-
-export default GObject.registerClass ({}, ClipShadowEffect)
+)
